Changed the format of sending data about product options in POST method to editProduct method in module model and controller files.
Added handling of product options to addProduct method in the module model file.

Fixed an error in module controller file, by putting a block of code into and if clause that checks for the presence of
product_id parameter in the incoming POST request.

Removed an error in module model file, by removing the SQL update to the nonexisting meta_title column.